# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Card Management API
  description: |
    REST API для управления банковскими картами,
    пользователями, запросами на блокировку
    и переводами между собственными картами.
  version: "1.0"
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/cards/{cardId}/block-requests:
    post:
      tags:
      - Заявки на блокировку карт
      summary: Создать заявку на блокировку карты
      description: |
        Создаёт заявку на блокировку карты,
        принадлежащей текущему пользователю.
      operationId: createBlockRequest
      parameters:
      - name: cardId
        in: path
        description: Идентификатор карты
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCardBlockRequest"
        required: true
      responses:
        "201":
          description: Заявка успешно создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardBlockRequestResponse"
        "400":
          description: Переданы некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: |
            Карта принадлежит другому пользователю
            или у пользователя недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: |
            Карта уже заблокирована
            или активная заявка уже существует
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
...
  /api/admin/users/{userId}/cards:
    post:
      tags:
      - Карты
      summary: Создать карту
      description: |
        Создаёт новую банковскую карту
        для указанного пользователя.
      operationId: createCard
      parameters:
      - name: userId
        in: path
        description: Идентификатор владельца карты
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "201":
          description: Карта успешно создана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
  /api/admin/cards/{cardId}/block:
    patch:
      tags:
      - Карты
      summary: Заблокировать карту
      description: Блокирует выбранную банковскую карту
      operationId: blockCard
      parameters:
      - name: cardId
        in: path
        description: Идентификатор карты
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Карта успешно заблокирована
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: |
            Карта уже заблокирована
            или срок её действия истёк
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
...
  /api/admin/block-requests/{requestId}/reject:
    patch:
      tags:
      - Заявки на блокировку карт
      summary: Отклонить заявку
      description: |
        Отклоняет заявку на блокировку.
        Статус карты при этом не изменяется.
      operationId: rejectBlockRequest
      parameters:
      - name: requestId
        in: path
        description: Идентификатор заявки
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessCardBlockRequest"
        required: true
      responses:
        "200":
          description: Заявка успешно отклонена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardBlockRequestResponse"
        "400":
          description: Переданы некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: |
            Заявка или администратор
            не найдены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: Заявка уже была обработана
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
  /api/admin/block-requests/{requestId}/approve:
    patch:
      tags:
      - Заявки на блокировку карт
      summary: Подтвердить заявку
      description: |
        Подтверждает заявку и блокирует
        связанную с ней карту.
      operationId: approveBlockRequest
      parameters:
      - name: requestId
        in: path
        description: Идентификатор заявки
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessCardBlockRequest"
        required: true
      responses:
        "200":
          description: |
            Заявка подтверждена,
            карта заблокирована
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardBlockRequestResponse"
        "400":
          description: Переданы некорректные данные
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: |
            Заявка, карта или администратор
            не найдены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "409":
          description: |
            Заявка уже была обработана
            или карта не может быть заблокирована
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
  /api/cards:
    get:
      tags:
      - Карты
      summary: Получить свои карты
      description: |
        Возвращает страницу банковских карт,
        принадлежащих текущему пользователю.

        Поддерживает поиск по последним четырём
        цифрам номера карты, фильтрацию по статусу,
        пагинацию и сортировку.
      operationId: getMyCards
      parameters:
      - name: lastFour
        in: query
        description: |
          Последние четыре цифры номера карты.
          Должны состоять ровно из четырёх цифр.
        required: false
        schema:
          type: string
        example: 3456
      - name: status
        in: query
        description: Статус карты
        required: false
        schema:
          type: string
          enum:
          - ACTIVE
          - BLOCKED
          - EXPIRED
        example: ACTIVE
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: Страница карт успешно получена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelCardResponse"
        "400":
          description: |
            Некорректные последние четыре цифры
            или значение статуса карты
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
  /api/cards/{cardId}:
    get:
      tags:
      - Карты
      summary: Получить свою карту
      description: |
        Возвращает карту только в том случае,
        если она принадлежит текущему пользователю.
      operationId: getMyCard
      parameters:
      - name: cardId
        in: path
        description: Идентификатор карты
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Карта успешно получена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Нет доступа к чужой карте
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
...
  /api/admin/cards:
    get:
      tags:
      - Карты
      summary: Получить все карты
      description: Возвращает страницу всех банковских карт
      operationId: getAllCards
      parameters:
      - name: page
        in: query
        description: Zero-based page index (0..N)
        required: false
        schema:
          type: integer
          default: 0
          minimum: 0
      - name: size
        in: query
        description: The size of the page to be returned
        required: false
        schema:
          type: integer
          default: 20
          minimum: 1
      - name: sort
        in: query
        description: "Sorting criteria in the format: property,(asc|desc). Default\
          \ sort order is ascending. Multiple sort criteria are supported."
        required: false
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: Страница карт успешно получена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PagedModelCardResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
  /api/admin/cards/{cardId}:
    get:
      tags:
      - Карты
      summary: Получить карту
      description: |
        Возвращает банковскую карту
        по её идентификатору.
      operationId: getCardById
      parameters:
      - name: cardId
        in: path
        description: Идентификатор карты
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "200":
          description: Карта успешно получена
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardResponse"
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
    delete:
      tags:
      - Карты
      summary: Удалить карту
      description: |
        Удаляет банковскую карту
        по её идентификатору.
      operationId: deleteCard
      parameters:
      - name: cardId
        in: path
        description: Идентификатор карты
        required: true
        schema:
          type: integer
          format: int64
        example: 1
      responses:
        "204":
          description: Карта успешно удалена
        "401":
          description: Пользователь не авторизован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "403":
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
        "404":
          description: Карта не найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorResponse"
      security:
      - bearerAuth: []
...
components:
  schemas:
...
    ApiErrorResponse:
      type: object
      description: Единый формат ошибки REST API
      properties:
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки
          example: 2026-07-28T19:34:52.426
        status:
          type: integer
          format: int32
          description: HTTP-статус ошибки
          example: 404
        error:
          type: string
          description: Наименование HTTP-ошибки
          example: Not Found
        message:
          type: string
          description: Сообщение для пользователя
          example: Карта с идентификатором 15 не найдена
        path:
          type: string
          description: Путь HTTP-запроса
          example: /api/cards/15
...
    CreateCardBlockRequest:
      type: object
      description: Данные для создания заявки на блокировку карты
      properties:
        reason:
          type: string
          description: Причина блокировки карты
          example: Карта была потеряна
          maxLength: 500
          minLength: 0
    CardBlockRequestResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cardId:
          type: integer
          format: int64
        status:
          type: string
          enum:
          - PENDING
          - APPROVED
          - REJECTED
        reason:
          type: string
        createdAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time
        processedById:
          type: integer
          format: int64
        adminComment:
          type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      description: JWT access-токен для авторизации
      scheme: bearer
      bearerFormat: JWT